pyproject.toml
[tool.poetry]
name = "journal-website"
version = "0.1.0"
description = ""
authors = ["DevMain1953 <alexanderbrenenburgh@gmail.com>"]
readme = "README.md"
packages = [{include = "journal_website"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2"
django-crispy-forms = "^2.0"
crispy-bootstrap4 = "^2022.1"
python-docx = "^0.8.11"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
